requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.41.2
torch==2.3.1
pypdf==4.2.0
fpdf==1.7.2
